补充状态修正

+++ engine/source/runtime/function/render/rhi/d3d12/d3d12_commandList.cpp
@@ -188,7 +188,7 @@ namespace RHI
     {
         std::vector<D3D12_RESOURCE_BARRIER> PendingResourceBarriersVector = ResolveResourceBarriers();
         
-        int PendingBatchNum = glm::ceil(PendingResourceBarriersVector.size() / NumBatches);
+        int PendingBatchNum = glm::ceil(PendingResourceBarriersVector.size() / (float)NumBatches);
         for (int i = 0; i < PendingBatchNum; i++)
         {
             int is = i * NumBatches;
